refactor(test/domain): 클래스별 테스트코드 분리
