+++ README.md
@@ -2,7 +2,7 @@
 
 블랙잭 미션 저장소
 
-- [ ] 카드를 섞는 기능
+- [x] 카드를 섞는 기능
 - [ ] 카드를 분배하는 기능 
     - 플레이어
       - [ ] 입력에 따라서 카드를 받을지말지 결정하는 기능
