+++ src/main/java/domain/BlackJackGame.java
@@ -1,0 +1,4 @@
+package domain;
+
+public class BlackJackGame {
+}
