+++ src/test/java/domain/ResultTest.java
@@ -1,0 +1,4 @@
+package domain;
+
+public class ResultTest {
+}
